chore(build): migrate from rollup to tsdown + add typecheck via tsgo
Replaces the rollup-based build pipeline with tsdown (rolldown) across
both packages. Adds a dedicated typecheck command powered by tsgo
(@typescript/native-preview, beta) and wires it into turbo + validate.

Build:
- packages/vidstack: rollup.config.ts -> tsdown.config.ts (NPM bundles,
  legacy CDN, modern CDN, plugins bundle, types bundle)
- packages/react: rollup.config.ts -> tsdown.config.ts (NPM bundles,
  types bundle with RSC directives and vidstack type resolution)
- Custom plugins ported: decorators, server-bundle stub, virtual env,
  rsc-directives, resolveGlobalTypes, rewriteVidstackImports,
  rewriteCDNChunks, copyAssets
- Types bundling now uses rolldown-plugin-dts (dtsInput mode), which
  also fixes the rollup-plugin-dts type-fest "Exported variable global
  is not defined" error

Typecheck:
- New `typecheck` script in both packages using `tsgo --noEmit`
- New `typecheck` turbo task with `^typecheck` dependency
- Root `typecheck` script and added to `validate` pipeline
- `declarations` now uses tsgo (~5x faster than tsc for .d.ts emit)

Removed deps: rollup, rollup-plugin-dts, @rollup/plugin-node-resolve,
magic-string (react), get-tsconfig
Added deps: tsdown, rolldown-plugin-dts, @typescript/native-preview

Other:
- tsconfig.build.json files: added rootDir (required by TS 6)
- turbo.json: added elements.json to analyze outputs (pre-existing bug)
- Outputs validated: 29/30 entry bundles export identical names; the
  30th (vidstack-remotion) has 4 fewer exports because tsdown no longer
  leaks internal symbols (ErrorBoundary, REMOTION_PROVIDER_ID, etc.)
  that were never in the explicit export list
