+++ packages/react/package.json
@@ -5,26 +5,29 @@
   "license": "MIT",
   "type": "module",
   "types": "src/index.ts",
-  "sideEffects": ["*.css"],
+  "sideEffects": [
+    "*.css"
+  ],
   "engines": {
     "node": ">=18"
   },
   "scripts": {
     "analyze": "maverick analyze \"src/**/*.tsx\" --framework react --project tsconfig.build.json",
     "dev": "pnpm clean && pnpm build:icons && pnpm declarations && pnpm watch",
-    "bundle": "rollup --config rollup.config.ts --configPlugin ../vidstack/build/rollup-ts.js",
+    "bundle": "tsdown --config-loader tsx",
     "build": "pnpm build:icons && pnpm bundle",
     "build:icons": "node build/build-icons.js",
-    "build:types": "pnpm run bundle --config-types",
-    "declarations": "tsc -p tsconfig.build.json",
+    "build:types": "pnpm run bundle --types",
+    "declarations": "tsgo -p tsconfig.build.json",
+    "typecheck": "tsgo -p tsconfig.json --noEmit",
     "sandbox": "node ../../.scripts/sandbox.js",
     "sandbox:build": "vite build sandbox",
     "types": "pnpm build:icons && pnpm declarations && pnpm build:types",
     "clean": "rimraf dist-npm",
     "format": "prettier src --write --loglevel warn",
     "link:styles": "ln -s ../vidstack/styles ./styles",
-    "watch": "pnpm watch:types & pnpm run bundle -w",
-    "watch:types": "pnpm run declarations -w & pnpm run build:types -w"
+    "watch": "pnpm watch:types & pnpm run bundle --watch",
+    "watch:types": "pnpm run declarations -w & pnpm run build:types --watch"
   },
   "peerDependencies": {
     "@types/react": "^18.0.0 || ^19.0.0",
@@ -36,27 +39,23 @@
   },
   "devDependencies": {
     "@maverick-js/cli": "0.43.2",
-    "@rollup/plugin-node-resolve": "^15.2.3",
     "@types/fs-extra": "^11.0.4",
     "@types/react": "^19.0.0",
     "@types/react-dom": "^19.0.0",
     "@vitejs/plugin-react": "^4.3.1",
-    "esbuild": "^0.23.0",
     "fs-extra": "^11.2.0",
-    "get-tsconfig": "^4.7.6",
-    "magic-string": "^0.30.11",
     "maverick.js": "0.43.2",
     "media-icons": "^1.1.5",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "remotion": "4.0.193",
     "rimraf": "^3.0.2",
-    "rollup": "^4.20.0",
-    "rollup-plugin-dts": "^6.1.1",
+    "rolldown-plugin-dts": "^0.25.1",
+    "tsdown": "^0.22.0",
     "tslib": "^2.6.3",
     "tsx": "^4.16.5",
-    "type-fest": "^3.13.1",
-    "typescript": "^5.6.0",
+    "type-fest": "^5.0.0",
+    "typescript": "^6.0.0",
     "vidstack": "workspace:*",
     "vite": "^5.3.5"
   },
