+++ packages/react/tsconfig.build.json
@@ -3,6 +3,7 @@
   "compilerOptions": {
     "declaration": true,
     "declarationDir": "types",
+    "rootDir": "..",
     "paths": {}
   },
   "include": ["src/**/*.ts", "src/**/*.tsx"],
