+++ packages/vidstack/package.json
@@ -18,13 +18,14 @@
   },
   "scripts": {
     "analyze": "maverick analyze \"src/**/*.ts\" -p tsconfig.build.json",
-    "build": "rollup --config rollup.config.ts --configPlugin ./build/rollup-ts.js",
-    "build:css": "pnpm run build --config-css",
-    "build:cdn": "rimraf dist-cdn && pnpm run build --config-cdn",
-    "build:plugins": "pnpm run build --config-plugins",
-    "build:types": "pnpm run build --config-types",
+    "build": "tsdown --config-loader tsx",
+    "build:css": "pnpm run build --css",
+    "build:cdn": "rimraf dist-cdn && pnpm run build --cdn",
+    "build:plugins": "pnpm run build --plugins",
+    "build:types": "pnpm run build --types",
     "dev": "pnpm clean && pnpm declarations && pnpm watch",
-    "declarations": "tsc -p tsconfig.build.json",
+    "declarations": "tsgo -p tsconfig.build.json",
+    "typecheck": "tsgo -p tsconfig.json --noEmit",
     "types": "pnpm declarations && pnpm run build:types",
     "clean": "rimraf dist-npm",
     "format": "prettier src --write --loglevel warn",
@@ -34,9 +35,9 @@
     "test:watch": "vitest --watch",
     "test:server": "SERVER=true vitest --run",
     "test:server:watch": "SERVER=true vitest --watch",
-    "watch": "pnpm run watch:types & pnpm run build -w",
-    "watch:cdn": "pnpm run build:cdn -w",
-    "watch:types": "pnpm run declarations -w & pnpm run build:types -w"
+    "watch": "pnpm run watch:types & pnpm run build --watch",
+    "watch:cdn": "pnpm run build:cdn --watch",
+    "watch:types": "pnpm run declarations -w & pnpm run build:types --watch"
   },
   "dependencies": {
     "@floating-ui/dom": "^1.6.10",
@@ -48,16 +49,14 @@
     "@maverick-js/cli": "0.43.2",
     "@open-wc/semantic-dom-diff": "^0.20.1",
     "@open-wc/testing-helpers": "^3.0.1",
-    "@rollup/plugin-node-resolve": "^15.2.3",
     "@types/fs-extra": "^11.0.4",
     "@types/fscreen": "^1.0.4",
     "autoprefixer": "^10.4.20",
     "compute-scroll-into-view": "^3.1.0",
     "dashjs": "^4.7.4",
     "es-module-lexer": "^1.5.4",
-    "esbuild": "^0.23.0",
+    "esbuild": "^0.28.0",
     "fscreen": "^1.2.0",
-    "get-tsconfig": "^4.7.6",
     "hls.js": "^1.5.14",
     "jassub": "^1.7.16",
     "jsdom": "^21.1.2",
@@ -68,16 +67,16 @@
     "media-icons": "^1.1.5",
     "postcss": "^8.4.41",
     "rimraf": "^3.0.2",
-    "rollup": "^4.20.0",
-    "rollup-plugin-dts": "^6.1.1",
+    "rolldown-plugin-dts": "^0.25.1",
     "rollup-pluginutils": "^2.8.2",
     "solid-js": "^1.8.19",
     "svelte": "^4.2.18",
     "tailwindcss": "^3.4.7",
+    "tsdown": "^0.22.0",
     "tslib": "^2.6.3",
     "tsx": "^4.16.5",
-    "type-fest": "^3.13.1",
-    "typescript": "^5.6.0",
+    "type-fest": "^5.0.0",
+    "typescript": "^6.0.0",
     "vite": "^5.3.5",
     "vitest": "^0.34.6",
     "vue": "^3.4.36"
